Checkbox + Post Card fixes + Signup/Login Metadas
- [X] Tried to fix the checkbox on mobile devices
- [X] Implemented some SEO optimization regarding metadatas in the Signup and Login pages
